Support multiple runners on a single instance with systemd templates
There are three systemd units:
1. `github-runner-install` - downloads and installs the latest version of the GitHub Runner. Wanted by `github-runner-config`.
2. `github-runner-config` - configures the recently downloaded GitHub Runner. Wanted by `github-runner`.
3. `github-runner` - Runs the recently installed and configured GitHub Runner.

Each of the above services relies on a `/etc/github-runner-env` file for the runner configuration environment, as well as
`/etc/environment` for the various hosted tools environment.

The AWS AMI also ships with systemd template units designed to support multiple runners. These look for environment files
in `/etc/github-runner-env-%i`, where `%i` is the identifier used when instantiating the service.